pyproject.toml

[tool.poetry]
name = "moviesoeiro"
version = "0.1.0"
description = "controle de filmes"
authors = ["mandresoeiro <seu@email.com>"]
packages = [{ include = "core" }]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
python-decouple = "^3.8"
mkdocs = "^1.5"
mkdocs-material = "^9.5"

[tool.poetry.group.dev.dependencies]
ipython = "^8.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.23"

